+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 1.63.0
 - Limit content length of requests when talking to the agent.
 - Fix: Handle non-pooled pg queries using the promise API correctly.